Finished chapter 6! We are rendering spheres now! Also, major refactoring and optimisations.
- added a new scene package
- added parallel rendering logic in new RenderSphere renderer
- migrated jagged arrays to mapped 1D arrays in Matrix
- moved from float to double all the way
- no journal entry this time :(

+++ src/main/java/org/schakalacka/java/raytracing/geometry/tracing/Ray.java
@@ -6,7 +6,7 @@ import org.schakalacka.java.raytracing.geometry.Tuple;
 public record Ray(Tuple origin, Tuple direction) {
 
     public Tuple position(double ticks) {
-        return origin.add(direction.mul((float) ticks));
+        return origin.add(direction.mul(ticks));
     }
 
     public Ray transform(Matrix translation) {
